change recover in TsFileManagement to not abstract method
