+++ server/src/main/java/org/apache/iotdb/db/engine/compaction/no/NoCompactionTsFileManagement.java
@@ -258,9 +258,6 @@ public class NoCompactionTsFileManagement extends TsFileManagement {
   }
 
   @Override
-  public void recover() {}
-
-  @Override
   public void forkCurrentFileList(long timePartition) {}
 
   @Override
